hazelcast-guides/event-driven-microservices-lab/event-generator/requirements.txt
confluent-kafka==2.3.0
hazelcast-python-client==5.3.0
